treatment dropdown fertig

+++ db/migrate/20140629223443_create_agentparams.rb
@@ -1,8 +1,8 @@
 class CreateAgentparams < ActiveRecord::Migration
   def change
     create_table :agentparams do |t|
-      t.float :infectious_dose
-      t.float :level_of_infectious_dose
+      t.decimal :infectious_dose
+      t.decimal :level_of_infectious_dose
       t.string :infected_species
       t.integer :agent_id
