+++ db/migrate/20140629224024_create_culdetecs.rb
@@ -3,7 +3,7 @@ class CreateCuldetecs < ActiveRecord::Migration
     create_table :culdetecs do |t|
       t.string :medium
       t.string :agar
-      t.float :incubation_duration
+      t.decimal :incubation_duration
       t.string :incubation_duration_unit
       t.string :incubation_temp
       t.integer :sample_volume_per_disc
