+++ db/migrate/20140629224103_create_results.rb
@@ -1,10 +1,10 @@
 class CreateResults < ActiveRecord::Migration
   def change
     create_table :results do |t|
-      t.integer :cfu_initial
+      t.decimal :cfu_initial
       t.string :cfu_initial_unit
       t.boolean :detectability
-      t.integer :cfu_detected
+      t.decimal :cfu_detected
       t.string :cfu_detected_unit
       t.text :r_comment
